eliminated api folder, moved into routes file
