--- server/api/comments.js
@@ -1,2 +1,0 @@
-// app.use('/api/comments', Comments);
-// const { Comments } = require('./api/comments');
